requirements.txt
python-telegram-bot==20.7
torch
pandas
numpy
scipy
matplotlib
requests
python-dotenv